refactor: move message package
