+++ common/types/message/message.go
