+++ common/types/message/message_test.go
