+++ roller/prover/mock.go
@@ -3,7 +3,7 @@
 package prover
 
 import (
-	"scroll-tech/common/message"
+	"scroll-tech/common/types/message"
 
 	"scroll-tech/roller/config"
 )
