+++ roller/prover/prover_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/scroll-tech/go-ethereum/core/types"
 	"github.com/stretchr/testify/assert"
 
-	"scroll-tech/common/message"
+	"scroll-tech/common/types/message"
 
 	"scroll-tech/roller/config"
 	"scroll-tech/roller/prover"
